+++ src/main/java/Deadline.java
@@ -1,5 +1,3 @@
-package duke.task;
-
 public class Deadline extends Task {
 
     protected String by;
