+++ src/main/java/Todo.java
@@ -1,5 +1,3 @@
-package duke.task;
-
 public class Todo extends Task {
 
     public Todo(String description) {
